# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Gerenciamento de Tarefas",
    "description": "API RESTful construída com Clean Architecture (Express + Sequelize + SQLite). Gerencia usuários e suas tarefas, aplicando a regra de negócio que limita cada usuário a no máximo 5 tarefas simultâneas com status EM_ANDAMENTO.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Usuários",
      "description": "Cadastro e consulta de usuários, os donos das tarefas."
    },
    {
      "name": "Tarefas",
      "description": "CRUD de tarefas e as transições de estado PENDENTE -> EM_ANDAMENTO -> CONCLUIDA."
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/usuarios/": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Lista todos os usuários",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista de usuários.",
            "schema": {
              "$ref": "#/definitions/ListaDeUsuarios"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Cria um novo usuário",
        "description": "Cadastra um usuário. O nome precisa ter no mínimo 3 caracteres e o e-mail precisa ser válido e único.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Dados do usuário: nome (mínimo 3 caracteres) e e-mail válido.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NovoUsuario"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado.",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          },
          "400": {
            "description": "Falha de validação estrutural do corpo (Zod), ou e-mail já cadastrado.",
            "schema": {
              "$ref": "#/definitions/ErroDeValidacao"
            }
          }
        }
      }
    },
    "/usuarios/{id}/tarefas": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Lista as tarefas de um usuário",
        "description": "Útil para conferir quantas tarefas do usuário estão EM_ANDAMENTO antes de tentar iniciar outra.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefas do usuário.",
            "schema": {
              "$ref": "#/definitions/ListaDeTarefas"
            }
          },
          "404": {
            "description": "Usuário não encontrado.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      }
    },
    "/tarefas/": {
      "get": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Lista todas as tarefas",
        "description": "Retorna todas as tarefas cadastradas, de todos os usuários, ordenadas por id.",
        "responses": {
          "200": {
            "description": "Lista de tarefas.",
            "schema": {
              "$ref": "#/definitions/ListaDeTarefas"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Cria uma nova tarefa",
        "description": "Cria uma tarefa vinculada a um usuário existente. Toda tarefa nasce com status PENDENTE.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Dados da tarefa: título (obrigatório), descrição (opcional) e o id do usuário dono (obrigatório).",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NovaTarefa"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Tarefa criada.",
            "schema": {
              "$ref": "#/definitions/Tarefa"
            }
          },
          "400": {
            "description": "Falha de validação estrutural do corpo (Zod): título ausente ou usuarioId inválido.",
            "schema": {
              "$ref": "#/definitions/ErroDeValidacao"
            }
          },
          "404": {
            "description": "Usuário informado não existe.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      }
    },
    "/tarefas/{id}": {
      "get": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Busca uma tarefa pelo id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id da tarefa"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa encontrada.",
            "schema": {
              "$ref": "#/definitions/Tarefa"
            }
          },
          "404": {
            "description": "Tarefa não encontrada.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Atualiza o texto de uma tarefa",
        "description": "Altera título e/ou descrição. O status não é alterado por esta rota.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id da tarefa"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Campos a alterar.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AtualizarTarefa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa atualizada.",
            "schema": {
              "$ref": "#/definitions/Tarefa"
            }
          },
          "400": {
            "description": "Título inválido.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          },
          "404": {
            "description": "Tarefa não encontrada.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Exclui uma tarefa",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id da tarefa"
          }
        ],
        "responses": {
          "204": {
            "description": "Tarefa excluída. Sem corpo na resposta."
          },
          "404": {
            "description": "Tarefa não encontrada.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      }
    },
    "/tarefas/{id}/iniciar": {
      "post": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Inicia uma tarefa (PENDENTE -> EM_ANDAMENTO)",
        "description": "REGRA DE NEGÓCIO: antes de iniciar, o serviço conta quantas tarefas do usuário dono já estão com status EM_ANDAMENTO. Se o total for igual ou superior a 5, a operação é barrada. Concluir uma tarefa libera uma vaga. O limite é contado por usuário, não globalmente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id da tarefa"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa iniciada com sucesso.",
            "schema": {
              "$ref": "#/definitions/TarefaEmAndamento"
            }
          },
          "400": {
            "description": "Retornado quando o usuário atinge o limite máximo de 5 tarefas com status EM_ANDAMENTO. Também cobre a transição inválida, quando a tarefa não está PENDENTE.",
            "schema": {
              "$ref": "#/definitions/ErroLimiteAtingido"
            }
          },
          "404": {
            "description": "Tarefa não encontrada.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      }
    },
    "/tarefas/{id}/concluir": {
      "post": {
        "tags": [
          "Tarefas"
        ],
        "summary": "Conclui uma tarefa (EM_ANDAMENTO -> CONCLUIDA)",
        "description": "Só é possível concluir uma tarefa que esteja EM_ANDAMENTO. Concluir libera uma vaga no limite de 5 tarefas em andamento do usuário.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Id da tarefa"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa concluída com sucesso.",
            "schema": {
              "$ref": "#/definitions/TarefaConcluida"
            }
          },
          "400": {
            "description": "Transição inválida: a tarefa não está EM_ANDAMENTO.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          },
          "404": {
            "description": "Tarefa não encontrada.",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nome": {
          "type": "string",
          "example": "Ana Souza"
        },
        "email": {
          "type": "string",
          "example": "ana.souza@exemplo.com"
        }
      }
    },
    "NovoUsuario": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Ana Souza"
        },
        "email": {
          "type": "string",
          "example": "ana.souza@exemplo.com"
        }
      },
      "required": [
        "nome",
        "email"
      ]
    },
    "Tarefa": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "titulo": {
          "type": "string",
          "example": "Modelar o banco de dados"
        },
        "descricao": {
          "type": "string",
          "example": "Definir as tabelas de usuários e tarefas"
        },
        "status": {
          "type": "string",
          "example": "PENDENTE"
        },
        "usuarioId": {
          "type": "number",
          "example": 1
        }
      }
    },
    "NovaTarefa": {
      "type": "object",
      "properties": {
        "titulo": {
          "type": "string",
          "example": "Modelar o banco de dados"
        },
        "descricao": {
          "type": "string",
          "example": "Definir as tabelas de usuários e tarefas"
        },
        "usuarioId": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "titulo",
        "usuarioId"
      ]
    },
    "AtualizarTarefa": {
      "type": "object",
      "properties": {
        "titulo": {
          "type": "string",
          "example": "Modelar o banco de dados - revisado"
        },
        "descricao": {
          "type": "string",
          "example": "Nova descrição da tarefa"
        }
      }
    },
    "ListaDeTarefas": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "titulo": {
            "type": "string",
            "example": "Modelar o banco de dados"
          },
          "descricao": {
            "type": "string",
            "example": "Definir as tabelas de usuários e tarefas"
          },
          "status": {
            "type": "string",
            "example": "PENDENTE"
          },
          "usuarioId": {
            "type": "number",
            "example": 1
          }
        }
      }
    },
    "ListaDeUsuarios": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "email": {
            "type": "string",
            "example": "ana.souza@exemplo.com"
          }
        }
      }
    },
    "TarefaEmAndamento": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "titulo": {
          "type": "string",
          "example": "Modelar o banco de dados"
        },
        "descricao": {
          "type": "string",
          "example": "Definir as tabelas de usuários e tarefas"
        },
        "status": {
          "type": "string",
          "example": "EM_ANDAMENTO"
        },
        "usuarioId": {
          "type": "number",
          "example": 1
        }
      }
    },
    "TarefaConcluida": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "titulo": {
          "type": "string",
          "example": "Modelar o banco de dados"
        },
        "descricao": {
          "type": "string",
          "example": "Definir as tabelas de usuários e tarefas"
        },
        "status": {
          "type": "string",
          "example": "CONCLUIDA"
        },
        "usuarioId": {
          "type": "number",
          "example": 1
        }
      }
    },
    "Erro": {
      "type": "object",
      "properties": {
        "erro": {
          "type": "string",
          "example": "Mensagem descrevendo o que impediu a operação."
        }
      }
    },
    "ErroLimiteAtingido": {
      "type": "object",
      "properties": {
        "erro": {
          "type": "string",
          "example": "Limite atingido: o usuário 1 já possui 5 tarefas com status EM_ANDAMENTO e o máximo permitido é 5. Conclua uma tarefa antes de iniciar outra."
        }
      }
    },
    "ErroDeValidacao": {
      "type": "object",
      "properties": {
        "erro": {
          "type": "string",
          "example": "Dados inválidos."
        },
        "detalhes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "campo": {
                "type": "string",
                "example": "titulo"
              },
              "mensagem": {
                "type": "string",
                "example": "O título da tarefa é obrigatório."
              }
            }
          }
        }
      }
    }
  }
}